Cleaned up routing and reused other lab components
