+++ client/src/components/imagine23/Main.js
@@ -4,7 +4,7 @@ import { navigate, Router } from "@reach/router";
 import { bindActionCreators } from "redux";
 import LandingPage from "./pages/landingPage";
 import MainInstructions from "./pages/mainInstructions";
-import Reading from "./pages/Reading";
+import { default as Reading } from "../body/Reading/Reading";
 import { default as ExerciseLab2 } from "../exercise/lab2/Main";
 import ExpressionStart from "./pages/ExpressionStart";
 import ExpressionExercise from "./pages/ExpressionExercise";
@@ -17,6 +17,7 @@ import Survey from "./pages/SurveyHandler";
 const { nanoid } = require("nanoid");
 import { PropTypes } from "prop-types";
 import useMainStateContext from "../../reducers/MainContext";
+import { default as Quiz } from "../quiz/components/QuizHandler";
 
 const mapStateToProps = (state) => ({
   state: state,
@@ -33,7 +34,7 @@ const Main = (props) => {
   const [count, setCount] = useState(0);
   const [userID, setUserID] = useState(null);
 
-  const [labId] = useState(2);
+  const labId = 2;
   const [isExperiential, setIsExperiential] = useState(false);
 
   const context = useMainStateContext();
@@ -154,7 +155,17 @@ const Main = (props) => {
             user={state.main.user}
             userID={userID}
             labID={labId}
-            imagine={true}
+            isImagine={isImagine}
+          />
+          <Quiz
+            path={'/Quiz'}
+            user={state.main.user}
+            labId={labId}
+            isImagine={isImagine}
+            hideCertificate={false}
+            submitData={() => { }}
+            isFinalQuiz={false}
+            userID={userID}
           />
           <Survey path={`/PostSurvey`} userID={userID} type="post" />
           <ExerciseEnd
@@ -172,7 +183,7 @@ const Main = (props) => {
 
 Main.propTypes = {
   actions: PropTypes.object,
-  state: PropTypes.string,
+  state: PropTypes.object,
   user: PropTypes.object,
   isImagine: PropTypes.bool,
 };
