+++ client/src/components/quiz/components/QuizHandler.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable no-unused-vars */
-import { React, useState } from "react";
+import { React, useEffect, useState } from "react";
 import { PropTypes } from "prop-types";
 import Quiz from "./Quiz";
 import Result from "./Result";
@@ -212,12 +212,22 @@ const QuizHandler = (props) => {
         );
       }
     } else {
-      props.submitData(
-        output,
-        props.user.userid,
-        props.labId,
-        (countCorrect / questionsTotal) * 100
-      );
+      if (props.isImagine) {
+        props.submitData(
+          output,
+          props.userID,
+          props.labId,
+          (countCorrect / questionsTotal) * 100
+        );
+      }
+      else {
+        props.submitData(
+          output,
+          props.user.userid,
+          props.labId,
+          (countCorrect / questionsTotal) * 100
+        );
+      }
     }
   }
 
@@ -259,9 +269,9 @@ const QuizHandler = (props) => {
       // checks to see if the set has the value in it
       !storageSet.has(answerValue)
         ? // adds it if it doesn't
-          storageSet.add(answerValue)
+        storageSet.add(answerValue)
         : // removes it if it does
-          storageSet.delete(answerValue);
+        storageSet.delete(answerValue);
       // assigns the updated set to the array
       tempAnswers[currentQuestionCursor] = storageSet;
     } else {
@@ -316,5 +326,6 @@ QuizHandler.propTypes = {
     userid: PropTypes.number,
   }),
   isImagine: PropTypes.bool,
+  userID: PropTypes.string
 };
 export default QuizHandler;
