+++ client/src/components/quiz/components/Result.js
@@ -4,6 +4,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import Certificate from "./Certificate";
+import { navigate } from "@reach/router";
 
 function Result(props) {
   function checkIfCorrect(answerIndex, questionIndex) {
@@ -119,29 +120,47 @@ function Result(props) {
     }
   }
 
+  const handleImagineSurvey = () => {
+    navigate("/Imagine/PostSurvey");
+  };
+
   return (
-    <div className="quiz container shadow">
-      <div className="result">
-        Results <strong>Score: {props.quizResult}</strong>
-        <br />
-        <div>
-          <table id="quizResults">
-            <tbody>
-              <tr>
-                <th>QUESTION</th>
-                <th>CORRECT ANSWERS</th>
-                <th>SELECTED ANSWERS</th>
-                <th>RESULTS</th>
-              </tr>
-              {renderTableData()}
-            </tbody>
-          </table>
-          {props.hideCertificate === false && (
-            <div style={{ marginTop: "50px" }}>
-              <Certificate quizResult={props.quizResult} lab={props.lab} />
-            </div>
-          )}
+    <div>
+      <div className="quiz container shadow">
+        <div className="result">
+          Results <strong>Score: {props.quizResult}</strong>
+          <br />
+          <div>
+            <table id="quizResults">
+              <tbody>
+                <tr>
+                  <th>QUESTION</th>
+                  <th>CORRECT ANSWERS</th>
+                  <th>SELECTED ANSWERS</th>
+                  <th>RESULTS</th>
+                </tr>
+                {renderTableData()}
+              </tbody>
+            </table>
+            {props.hideCertificate === false && (
+              <div style={{ marginTop: "50px" }}>
+                <Certificate quizResult={props.quizResult} lab={props.lab} />
+              </div>
+            )}
+          </div>
         </div>
+      </div>
+      <div className=" d-flex flex-column justify-content-center mt-3">
+        {props.isImagine ? (
+          <button
+            className="btn btn-primary btn-xl text-uppercase  next"
+            onClick={handleImagineSurvey}
+          >
+            Continue to Post-Survey
+          </button>
+        ) : (
+          <></>
+        )}
       </div>
     </div>
   );
